Split src/resources/projects into module dir
